Staff Page, New CSS, Modified Pages, Basic Page
Made a page for the Staff.
Modified the Index page.
Replaced card.css with another snippet which was more fitting.
Made a basic.html page with just Navbar, Footer and a Placeholder
paragraph in center, to be used for basic structure of all future pages.
Added a new footer and footer.css
Added carousel to index.html and made carousel.css